+++ x-pack/plugin/old-lucene-versions/src/main/java/org/elasticsearch/xpack/lucene/bwc/codecs/lucene54/Lucene54DocValuesConsumer.java
@@ -343,7 +343,7 @@ final class Lucene54DocValuesConsumer extends DocValuesConsumer implements Close
                 bits = 0;
             }
             if (v != null) {
-                bits |= 1 << (count & 7);
+                bits |= (byte) (1 << (count & 7));
             }
             count++;
         }
